+++ minecraft/src/net/minecraft/client/gui/layouts/FrameLayout.java
@@ -98,7 +98,7 @@ public class FrameLayout extends AbstractLayout {
 	}
 
 	public static void alignInDimension(int i, int j, int k, Consumer<Integer> consumer, float f) {
-		int l = Mth.lerp(f, 0, j - k);
+		int l = (int)Mth.lerp(f, 0.0F, (float)(j - k));
 		consumer.accept(i + l);
 	}
